Restructures the repository for better organization
Refactors the project by moving files and directories into a more organized structure.
This change enhances maintainability and clarity by grouping related components logically within the `topicer` directory.
Specifically, moves core modules like data handling and tagging into their respective subdirectories.

Removes unnecessary files.
