+++ config.ini
@@ -1,3 +1,5 @@
+# --- SSH TUNNEL CONFIG ---
+
 [SSH]
 # Username on the remote server
 user = xjuric31
